refactor(api): move isInQuietHours out of the Firebase notifications module (#3901)
